o1 preview. 15 ekim oncesi pazar gunu: SubscriptionModel ve ona uygun diğer claslar güncellendi admin tarafı eklemeler vs.

+++ lib/managers/appointment_manager.dart
@@ -1,6 +1,9 @@
+// appointment_manager.dart
+
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
-import '../commons/common.dart';
+
 import '../commons/logger.dart';
 import '../commons/userclass.dart';
 
@@ -55,14 +58,22 @@ class AppointmentManager extends ChangeNotifier {
     return availableSlots;
   }
 
-  Future<List<AppointmentModel>> fetchCurrentUserAppointments(
-      String userId) async {
+  Future<List<AppointmentModel>> fetchCurrentUserAppointments() async {
+    final user = FirebaseAuth.instance.currentUser;
+    if (user == null) {
+      logger.err('User not authenticated.');
+      return [];
+    }
+
     final collectionRef = FirebaseFirestore.instance
         .collection('users')
-        .doc(userId)
+        .doc(user.uid)
         .collection('appointments');
 
-    final querySnapshot = await collectionRef.get();
+    final querySnapshot = await collectionRef
+        .orderBy('appointmentDateTime', descending: false)
+        .get();
+
     List<AppointmentModel> userAppointments = [];
 
     for (var doc in querySnapshot.docs) {
@@ -81,7 +92,7 @@ class AppointmentManager extends ChangeNotifier {
     if (selectedDate != null) {
       DateTime startOfDay =
       DateTime(selectedDate.year, selectedDate.month, selectedDate.day);
-      DateTime endOfDay = startOfDay.add(Duration(days: 1));
+      DateTime endOfDay = startOfDay.add(const Duration(days: 1));
 
       query = query.where('appointmentDateTime',
           isGreaterThanOrEqualTo: Timestamp.fromDate(startOfDay),
@@ -104,27 +115,81 @@ class AppointmentManager extends ChangeNotifier {
     }
   }
 
-  Future<void> bookAppointment(AppointmentModel appointment) async {
-    bool isAvailable = isTimeSlotAvailable(
-      appointment.appointmentDateTime,
-      TimeOfDay.fromDateTime(appointment.appointmentDateTime),
+  Future<void> bookAppointmentForCurrentUser() async {
+    final user = FirebaseAuth.instance.currentUser;
+    if (user == null) {
+      logger.err('User not authenticated.');
+      throw Exception('User not authenticated.');
+    }
+
+    final subscription = await getCurrentSubscription();
+    if (subscription == null) {
+      throw Exception('No active subscription found.');
+    }
+
+    if (_selectedTimeNotifier.value == null) {
+      throw Exception('Please select a time slot.');
+    }
+
+    DateTime appointmentDateTime = DateTime(
+      _selectedDate.year,
+      _selectedDate.month,
+      _selectedDate.day,
+      _selectedTimeNotifier.value!.hour,
+      _selectedTimeNotifier.value!.minute,
     );
-    if (isAvailable) {
-      try {
-        await FirebaseFirestore.instance
-            .collection('users')
-            .doc(appointment.userId)
-            .collection('appointments')
-            .doc(appointment.appointmentId)
-            .set(appointment.toMap());
-        await fetchAppointments();
-        notifyListeners();
-      } catch (error) {
-        logger.err('Error booking appointment: {}', [error]);
-        throw Exception('Error booking appointment.');
-      }
-    } else {
+
+    bool isAvailable = isTimeSlotAvailable(
+      _selectedDate,
+      _selectedTimeNotifier.value!,
+    );
+
+    if (!isAvailable) {
       throw Exception('Selected time slot is not available.');
+    }
+
+    try {
+      final appointmentId = FirebaseFirestore.instance
+          .collection('users')
+          .doc(user.uid)
+          .collection('appointments')
+          .doc()
+          .id;
+
+      AppointmentModel appointment = AppointmentModel(
+        appointmentId: appointmentId,
+        userId: user.uid,
+        subscriptionId: subscription.subscriptionId,
+        meetingType: _meetingTypeNotifier.value,
+        appointmentDateTime: appointmentDateTime,
+        status: 'scheduled',
+        createdAt: DateTime.now(),
+      );
+
+      await FirebaseFirestore.instance
+          .collection('users')
+          .doc(user.uid)
+          .collection('appointments')
+          .doc(appointmentId)
+          .set(appointment.toMap());
+
+      // Update the subscription's meetingsRemaining
+      //BEN: NO, ADMIN ONAYIYLA OLMALI. TODO
+      // subscription.meetingsRemaining -= 1;
+      // await FirebaseFirestore.instance
+      //     .collection('users')
+      //     .doc(user.uid)
+      //     .collection('subscriptions')
+      //     .doc(subscription.subscriptionId)
+      //     .update({
+      //   'meetingsRemaining': subscription.meetingsRemaining,
+      // });
+
+      await fetchAppointments();
+      notifyListeners();
+    } catch (error) {
+      logger.err('Error booking appointment: {}', [error]);
+      throw Exception('Error booking appointment.');
     }
   }
 
@@ -139,16 +204,37 @@ class AppointmentManager extends ChangeNotifier {
     return true;
   }
 
-  Future<void> cancelAppointment(String userId, String appointmentId) async {
+  Future<void> cancelAppointment(String appointmentId) async {
+    final user = FirebaseAuth.instance.currentUser;
+    if (user == null) {
+      logger.err('User not authenticated.');
+      throw Exception('User not authenticated.');
+    }
+
     try {
       await FirebaseFirestore.instance
           .collection('users')
-          .doc(userId)
+          .doc(user.uid)
           .collection('appointments')
           .doc(appointmentId)
           .delete();
 
       logger.info('Appointment with id={} has been canceled.', [appointmentId]);
+
+      // Optionally, you might want to update the subscription's meetingsRemaining
+      //BEN: NO, ADMIN TARAFININ GUNCELLEMESIYLE OLMALI. TODO
+      // final subscription = await getCurrentSubscription();
+      // if (subscription != null) {
+      //   subscription.meetingsRemaining += 1;
+      //   await FirebaseFirestore.instance
+      //       .collection('users')
+      //       .doc(user.uid)
+      //       .collection('subscriptions')
+      //       .doc(subscription.subscriptionId)
+      //       .update({
+      //     'meetingsRemaining': subscription.meetingsRemaining,
+      //   });
+      // }
 
       await fetchAppointments();
       notifyListeners();
@@ -157,5 +243,29 @@ class AppointmentManager extends ChangeNotifier {
           [appointmentId, error]);
       throw Exception('Error canceling appointment.');
     }
+  }
+
+  Future<SubscriptionModel?> getCurrentSubscription() async {
+    final user = FirebaseAuth.instance.currentUser;
+    if (user == null) {
+      logger.err('User not authenticated.');
+      return null;
+    }
+    try {
+      final snapshot = await FirebaseFirestore.instance
+          .collection('users')
+          .doc(user.uid)
+          .collection('subscriptions')
+          .where('status', isEqualTo: 'active')
+          .limit(1)
+          .get();
+
+      if (snapshot.docs.isNotEmpty) {
+        return SubscriptionModel.fromDocument(snapshot.docs.first);
+      }
+    } catch (e) {
+      logger.err('Error fetching current subscription: {}', [e]);
+    }
+    return null;
   }
 }
